Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
-    float x;
+    int x;
     scanf("%f", &x);
-    printf("You entered: %.2f", x);
+    printf("You entered: %f", x);
     return 0;
 }
